use homebrew on mac
